DatingAppPythonBackend/requirements.txt
Flask_SQLAlchemy==2.4.1
messagebird==1.5.0
Flask==1.1.1
Flask_Cors==3.0.8
SQLAlchemy==1.3.12
imgurpython==1.1.7
PyJWT==1.7.1
psycopg2==2.7.3.1
boto3==1.13.8
google-cloud==0.34.0
google-cloud-vision==1.0.0
gensim==3.8.3
